# docs/swagger.yaml
basePath: /api
definitions:
  entity.DontStarveCluster:
    properties:
      cluster_description:
        type: string
      cluster_name:
        type: string
      cluster_password:
        type: string
      created_at:
        type: string
      id:
        type: string
      max_players:
        type: string
      mods:
        items:
          $ref: '#/definitions/entity.DontStarveModInCluster'
        type: array
      pvp:
        type: string
      updated_at:
        type: string
      worlds:
        items:
          $ref: '#/definitions/entity.DontStarveWorld'
        type: array
    type: object
  entity.DontStarveConfig:
    properties:
      cluster_token:
        type: string
    type: object
...
  entity.DontStarveMod:
    properties:
      config:
        type: string
      created_at:
        type: string
      id:
        type: string
      mod_id:
        type: string
      name:
        type: string
      remark:
        type: string
      updated_at:
        type: string
    type: object
  entity.DontStarveModInCluster:
    properties:
      created_at:
        type: string
      current_config:
        type: string
      dont_starve_cluster_id:
        type: string
      id:
        type: string
      mod:
        $ref: '#/definitions/entity.DontStarveMod'
      mod_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.DontStarveWorld:
    properties:
      created_at:
        type: string
      dont_starve_cluster_id:
        type: string
      id:
        type: string
      server_config:
        type: string
      type:
        type: string
      updated_at:
        type: string
      world_override:
        type: string
    type: object
...
  http.response:
    type: object
host: 0.0.0.0:5800
info:
  contact: {}
  description: Aurora Admin API
  title: Aurora Admin
  version: "1.0"
paths:
  /api/dontstarve/cluster/:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DontStarveCluster'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 创建存档
      tags:
      - dontstarve_cluster
  /api/dontstarve/cluster/{id}:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 删除存档
      tags:
      - dontstarve_cluster
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DontStarveCluster'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 获取存档信息
      tags:
      - dontstarve_cluster
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.DontStarveCluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DontStarveCluster'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 更新存档
      tags:
      - dontstarve_cluster
  /api/dontstarve/clusters:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.DontStarveCluster'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 获取存档列表
      tags:
      - dontstarve_cluster
  /api/dontstarve/config:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DontStarveConfig'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 查看配置
      tags:
      - dontstarve_config
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DontStarveConfig'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 更新配置
      tags:
      - dontstarve_config
  /api/dontstarve/console/announce/{msg}:
    post:
      consumes:
      - application/json
      parameters:
      - description: msg
        in: path
        name: msg
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 全服宣告
      tags:
      - dontstarve_console
  /api/dontstarve/console/player:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 获取当前玩家
      tags:
      - dontstarve_console
  /api/dontstarve/console/regenerate:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 重新生成世界
      tags:
      - dontstarve_console
  /api/dontstarve/console/rollback{days}:
    post:
      consumes:
      - application/json
      parameters:
      - description: msg
        in: path
        name: times
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 回滚
      tags:
      - dontstarve_console
  /api/dontstarve/control/{action}:
    post:
      consumes:
      - application/json
      parameters:
      - description: '[ start | stop | restart | update | install ]'
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 服务器控制
      tags:
      - dontstarve_control
...
  /api/dontstarve/mod:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 删除 Mods
      tags:
      - dontstarve_mod
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.DontStarveMod'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 查看 Mods
      tags:
      - dontstarve_mod
    put:
      consumes:
      - application/json
      parameters:
      - description: '[ true | false ]'
        in: query
        name: update_mod_info
        type: string
      - description: body
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/entity.DontStarveMod'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 更新 Mods
      tags:
      - dontstarve_mod
  /api/dontstarve/mod/mod_id:
    post:
      consumes:
      - text/plain
      parameters:
      - description: 多个 Mod ID 一行一个
        in: body
        name: request
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: 通过 Mod ID 添加 Mods
      tags:
      - dontstarve_mod
...
swagger: "2.0"